@@ -1,6 +1,7 @@
 import { v4 as uuidV4 } from "uuid";
 import {Column, CreateDateColumn, Entity, PrimaryColumn} from "typeorm";
 
+
 @Entity("categories")
 class Category {
     @PrimaryColumn()
@@ -11,7 +12,7 @@ class Category {
 
     @Column()
     description: string;
-    
+
     @CreateDateColumn()
     created_at: Date;
 
