@@ -1,14 +1,15 @@
 import { Response, Request } from "express";
-
+import { container } from "tsyringe";
 import { CreateCategoryUseCase } from "./CreateCategoryUseCase";
 
 class CreateCategoryController {
-    constructor(private createCategoryUseCase: CreateCategoryUseCase) {}
 
     async handle(request: Request, response: Response): Promise<Response> {
         const { name, description } = request.body;
 
-        await this.createCategoryUseCase.execute({ name, description });
+        const createCategoryUseCase = container.resolve(CreateCategoryUseCase);
+
+        await createCategoryUseCase.execute({ name, description });
 
         return response.status(201).send();
     }
