@@ -1,7 +1,7 @@
 import { Router } from "express";
 import multer from "multer";
 
-import  createCategoryController  from "../modules/car/useCases/createCategory";
+import { CreateCategoryController } from "../modules/car/useCases/createCategory/CreateCategoryController";
 import { importCategoryController } from "../modules/car/useCases/importCategory";
 import { listCategoriesController } from "../modules/car/useCases/listCategories.ts";
 
@@ -9,9 +9,9 @@ const categoriesRoutes = Router();
 
 const upload = multer({ dest: "./tmp" });
 
-categoriesRoutes.post("/", (request, response) => {
-    return createCategoryController().handle(request, response);
-});
+const createCategoryController = new CreateCategoryController();
+
+categoriesRoutes.post("/", createCategoryController.handle);
 
 categoriesRoutes.get("/", (request, response) => {
     return listCategoriesController.handle(request, response);
