@@ -1,10 +1,13 @@
 import express from "express";
 import swaggerUi from "swagger-ui-express";
 
+import "./database";
+
+import "./shared/container";
+
 import { router } from "./routes";
 import swaggerFile from "./swagger.json";
 
-import "./database";
 
 const app = express();
 
